[FEAT] 대시보드 API명세서에 맞게 수정
